android dev: minor ui change

+++ dcc-credit-android/dccchainservice/src/main/java/io/wexchain/dccchainservice/domain/EcoBonus.kt
@@ -1,12 +1,12 @@
 package io.wexchain.dccchainservice.domain
 import com.google.gson.annotations.SerializedName
-import java.math.BigDecimal
+import java.math.BigInteger
 
 
 data class EcoBonus(
     @SerializedName("id") val id: Long,
     @SerializedName("receiverAddress") val receiverAddress: String,
-    @SerializedName("amount") val amount: BigDecimal?,
+    @SerializedName("amount") val amount: BigInteger?,
     @SerializedName("memo") val memo: String?,
     @SerializedName("createdTime") val createdTime: Long
 )
